% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExportToVorteks.R
\name{ExportDataframeToVorteks}
\alias{ExportDataframeToVorteks}
\title{Export dataframe to Vorteks Visualization Environment}
\usage{
ExportDataframeToVorteks(df, filename = "dataframe_export")
}
\arguments{
\item{df}{dataframe variable}

\item{filename}{name of exported file variable}
}
\description{
Export dataframe to Vorteks Visualization Environment
}
