% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExportToVorteks.R
\name{ExportDataframeToVPE}
\alias{ExportDataframeToVPE}
\title{Export dataframe to Vorteks Processing Environment (VPE)}
\usage{
ExportDataframeToVPE(df, filename = "dataframe_export")
}
\arguments{
\item{df}{dataframe variable}

\item{filename}{name of exported file variable}
}
\description{
Export dataframe to Vorteks Processing Environment (VPE)
}
